silence openssl output

+++ make-tls-certs.sh
@@ -1,5 +1,4 @@
-#!/usr/bin/env bash
-set -euo pipefail
+#!/usr/bin/env sh
 
 # Detect OpenSSL flavor and set correct flag
 OPENSSL_REQ_NOENC_FLAG="-noenc"
@@ -18,19 +17,30 @@ fi
 # If certs/rootCA.key does NOT exist, create it
 if [ ! -f certs/root/rootCA.key ]
 then
+    echo "Generating key for rootCA ..."
     # Create the root key
-    openssl genrsa -out certs/root/rootCA.key 4096
+    openssl genrsa -out certs/root/rootCA.key 4096 >/dev/null 2>&1
+    echo "    certs/root/rootCA.key"
+    echo "    Done."
+else
+    echo "Detected key for rootCA at certs/root/rootCA.key. Skipping new key generation..."
 fi
 
 # If certs/rootCA.crt does NOT exist, create it
 if [ ! -f certs/root/rootCA.crt ]
 then
+    echo "Generating cert for rootCA ..."
     # Generate the Root Certificate.
     openssl req -x509 -sha256 -new $OPENSSL_REQ_NOENC_FLAG \
         -key certs/root/rootCA.key \
         -out certs/root/rootCA.crt \
         -subj "/C=$COUNTRY/ST=$STATE/L=$LOCALITY/O=$ORGANIZATION CA/OU=$ORGANIZATIONAL_UNIT/CN=$ROOT_CN" \
-        -days 800
+        -days 800 \
+        >/dev/null 2>&1
+    echo "    certs/root/rootCA.crt"
+    echo "    Done."
+else
+    echo "Detected cert for rootCA at certs/root/rootCA.crt. Skipping new cert generation..."
 fi
 
 # If certs/servers does NOT exist, create it
@@ -76,7 +86,8 @@ get_all_hosts() {
 
 # Function to gather all system hostnames + IPs (macOS + Linux)
 generate_server_cert_key() {
-      local server="$1"
+    server="$1"
+    echo "Generating cert/key for $server ..."
 
     # If certs/servers/$server does NOT exist, create it
     if [ ! -d "certs/servers/$server" ]
@@ -89,17 +100,25 @@ generate_server_cert_key() {
     then
         openssl genpkey -algorithm RSA \
             -out "certs/servers/$server/key.pem" \
-            -pkeyopt rsa_keygen_bits:4096
+            -pkeyopt rsa_keygen_bits:4096 \
+            >/dev/null 2>&1
+        echo "    Generated key at certs/servers/$server/key.pem"
+    else
+        echo "    Detected key at certs/servers/$server/key.pem. Skipping new key generation..."
     fi
 
-    # Generate the Certificate Signing Request (CSR)
-    openssl req -new \
-        -key "certs/servers/$server/key.pem" \
-        -out "certs/servers/$server.csr" \
-        -subj "/C=$COUNTRY/ST=$STATE/L=$LOCALITY/O=$ORGANIZATION certificate/OU=$ORGANIZATIONAL_UNIT/CN=$server"
+    # Create the certificate if it doesn't exist
+    if [ ! -f "certs/servers/$server/cert.pem" ]
+    then
+        # Generate the Certificate Signing Request (CSR)
+        openssl req -new \
+            -key "certs/servers/$server/key.pem" \
+            -out "certs/servers/$server.csr" \
+            -subj "/C=$COUNTRY/ST=$STATE/L=$LOCALITY/O=$ORGANIZATION certificate/OU=$ORGANIZATIONAL_UNIT/CN=$server" \
+            >/dev/null 2>&1
 
-    # Configure extensions so browsers don't yell
-    cat > "certs/servers/$server.ext" << EOF
+        # Configure extensions so browsers don't yell
+        cat > "certs/servers/$server.ext" << EOF
 authorityKeyIdentifier=keyid,issuer
 basicConstraints=CA:FALSE
 keyUsage = digitalSignature, nonRepudiation, keyEncipherment, dataEncipherment
@@ -108,22 +127,24 @@ subjectAltName = @alt_names
 DNS.1 = $server
 EOF
 
-    # Finally create the certificate for the server and sign it using CA
-    openssl x509 -req \
-        -in "certs/servers/$server.csr" \
-        -CA certs/root/rootCA.crt \
-        -CAkey certs/root/rootCA.key \
-        -CAcreateserial \
-        -out "certs/servers/$server/cert.pem" \
-        -days 825 -sha256 \
-        -extfile "certs/servers/$server.ext"
+        # Finally create the certificate for the server and sign it using CA
+        openssl x509 -req \
+            -in "certs/servers/$server.csr" \
+            -CA certs/root/rootCA.crt \
+            -CAkey certs/root/rootCA.key \
+            -CAcreateserial \
+            -out "certs/servers/$server/cert.pem" \
+            -days 825 -sha256 \
+            -extfile "certs/servers/$server.ext" \
+            >/dev/null 2>&1
 
-    # Remove unnecessary files: Certificate Signing Request (CSR).
-    rm "certs/servers/$server.csr" "certs/servers/$server.ext" certs/root/rootCA.srl
+        # Remove unnecessary files: Certificate Signing Request (CSR).
+        rm "certs/servers/$server.csr" "certs/servers/$server.ext" "certs/root/rootCA.srl"
 
-    # Verify the newly created certificate
-    echo "Created a certificate and a key for '$server' at certs/servers/$server/cert.pem and certs/servers/$server/key.pem"
-    openssl x509 -in "certs/servers/$server/cert.pem" -text -noout
+        echo "    Generated cert at certs/servers/$server/cert.pem"
+    else
+        echo "    Detected cert at certs/servers/$server/cert.pem. Skipping new cert generation..."
+    fi
 }
